Eng 12089 Validate BIGINT, DOUBLE and TIMESTAMP constant values from HSQL. (#4351)
If we parse a constant VoltXML expression whose string value cannot be a
member of the expression's type, we should throw an exception, even if
the string value has been extracted as a parameter.

https://issues.voltdb.com/browse/ENG-12089
(Note: The ticket above complains of multiple errors.  This is just one of them.  See ENG-12076 and ENG-12116 for the root causes of other problems exposed in this problem report.)