ENG-11662: Relieve internal adapter lock contention. (#4332)
* ENG-11662: Relieve internal adapter lock contention.

Instead of using a global internal adapter for all importer
transactions, create one adapter for each partition lazily. This
relieves the contention on the ClientInterfaceHandleManager because the
usage will be bound to the same thread for each partition.

The adapters are created when partition leaders are elected so that it
can also pick up elastically added partitions.

* ENG-11662: Change importer back pressure.

Use a simple permit system for back pressure. The old back pressure
mechanism tend to sleep for 256ms for each importer thread most of the
time, which was too long and it created a lot of burstiness.

The permit is enforced in the per partition internal adapters. Each
partition has a semaphore to tracker the permits, also avoids contention
on a global semaphore. Each partition has 500 permits by default. This
value can be changed by setting the
INTERNAL_MAX_PENDING_TRANSACTION_PER_PARTITION Java system property.