Eng 12025 Kafka import benchmark (#4309)
* ENG-12025 Initial import benchmark

* ENG-12025 progress

* ENG-12025 progress

* ENG-12025 save

* ENG-12025 getting stats...

* ENG-12025 Rearrange stats; revert Producer changes.

* ENG-12025 Change producer "acks" to speed it up.

* ENG-12025 Remove leftover debug echo.

* ENG-12025 space & tab correction.

* ENG-11662: Relieve internal adapter lock contention.

Instead of using a global internal adapter for all importer
transactions, create one adapter for each partition lazily. This
relieves the contention on the ClientInterfaceHandleManager because the
usage will be bound to the same thread for each partition.

The adapters are created when partition leaders are elected so that it
can also pick up elastically added partitions.

* ENG-11662: Change importer back pressure.

Use a simple permit system for back pressure. The old back pressure
mechanism tend to sleep for 256ms for each importer thread most of the
time, which was too long and it created a lot of burstiness.

* ENG-12025 updates to track ENG-11662 changes.