sql-grammar.txt: 3 major changes: added Window functions (RANK, etc.), and associated syntax; expanded the use of aggregate functions, used as Window functions and otherwise; expanded the use of sub-queries, scalar and otherwise. And multiple minor changes: added very occasional use of nonexistent table, view, and column names; added symbols for byte columns and expressions, and use them as args for the REPEAT function, to avoid ENG-12118; improved UPDATE statements, to get the column types right most of the time; expanded boolean expressions, to include multiple AND or OR, optionally using parentheses; fixed USING clauses, by adding parentheses; added EXISTS and IN clauses; removed the nonexistent % operator (SQL has the MOD function, but no remainder op); removed the use of CONCAT with only one argument, which does not work; fixed use of the TRIM function, by adding single-quotes and fixing an unmatched '['; added the beginnings of varbinary, point, and polygon expressions (though more work is needed in this area); and other minor tweaks, e.g. to likelihoods of different grammar options.
sql_grammar_generator.py: 
added the ability to specify SQL statements containing multiple phrases
(e.g., 'INSERT' and 'SELECT') to be echoed to a file (with their
results), and optionally, with the grammar symbols that were used to
create them, and the ability to count the number and percent of such
statements that were valid versus invalid (all of which helps to debug
changes to the grammar file);
added (to the summary file) counts of the number of certain Java
Exceptions (e.g. NPE's) found in the VoltDB log or console output; 
and tweaked various comments and debug print.