ENG-11957, fix a wrong use of hashCode() in HSQLDB parser, which may lead to hash collision in some circustances.
When HSQLDB parse the sql expression, it uses a hash map to cache expression ids it has already computed. But it
uses the hash code of object instead of the object itself as the key, which is not guaranteed to be unique for different
instances. In some rare cases, this will cause hash key collision which let the SQL parser returns incorrect columns for
values to be inserted into the table.

Change-Id: I35e6cda7ea0b93f1e63a7f20a8d478b0298995a5